requirements.txt
cbpro~=1.1.4
mysql-connector-python~=8.0.28